Reduce number of build tasks
This moves some of the logic from YML to ps1 scripts where it can be reproduced locally and therefore is more diagnosable and maintainable.

It also adds an init script to the repo root which installs whatever .NET Core SDK and runtimes are required to build and run tests in this repo.