Remove formatter instance cache from TypelessFormatter
The `TypelessFormatter` would get its formatter objects from `options.Resolver` argument, which can of course vary from call to call. Yet this `TypelessFormatter` was stored by `TypelessObjectResolver` which in turn is stored in the static, shared `TypelessContractlessStandardResolver`. This is a big problem because that means over time using the `TypelessContractlessStandardResolver` with different resolvers passed through `options` could lead to the `TypelessFormatter` using the wrong formatters for a given serialization.

Now that these instances are truly safe to share, I made a couple more singletons.